+++ core/src/main/java/io/jenkins/pluginhealth/scoring/service/ScoreService.java
@@ -86,7 +86,7 @@ public class ScoreService {
     }
 
     @Transactional
-    public long deleteOldScores() {
+    public int deleteOldScores() {
         return repository.deleteOldScoreFromPlugin();
     }
